Refactoring of io utils, no change in imports
