+++ mlcolvar/utils/io/__init__.py
@@ -1,0 +1,18 @@
+"""Input/output functions."""
+try:
+    import pandas as pd
+except ImportError as e:
+    raise ImportError(
+        "pandas is required to use the i/o utils (mlcolvar.utils.io)\n", e
+    )
+
+__all__ = ["load_dataframe", 
+           "plumed_to_pandas", 
+           "create_dataset_from_files",
+           "create_dataset_from_configurations", 
+           "create_dataset_from_trajectories",
+           "create_pdb_from_xyz"]
+
+# not imported by default as they depend on optional libraries (pandas, scikit-learn or KDEpy)
+from .colvar import *
+from .traj import *
